requirements.txt
pyfiglet
pyttsx3
python-barcode